Use collab for new GitHub Organization permissions
Because of the new [GitHub Organization
permissions](https://github.com/blog/2020-improved-organization-permissions)
users don't have to be on a team in order to be apart of an
organization.

This means that for individual assignments we make the user an outside
collaborator, in lieu of them being on their own personal team.

We also add users to a team only when they are apart of group
assignment.