Create an repo permission info box
This alerts the users that they should change their Default repository
permission level to none in order to make sure that students can only
see the assignments they are assigned to.

Once the user has click the close button a cookie is set so that they
won't see it again.